⚡️ ::  memberMapper 디렉터리 수정 - 소연
⚡️ ::  memberMapper 디렉터리 수정 - 소연
